+++ graph_db/repository/repository_interfaces.py
