--- src/cli/cli_controller.py
@@ -1,57 +1,0 @@
-"""This module provides the CLIController class, the entry point for the CLI, handling user commands."""
-
-import argparse
-from src.core.application_server import ApplicationServer
-
-
-class CLIController:
-    """
-    CLIController handles the parsing of command-line arguments and triggers actions within the application.
-    """
-
-    def __init__(self):
-        self.parser = argparse.ArgumentParser(description="CLI for ETL Process")
-        self.setup_arguments()
-        self.app_server = ApplicationServer()
-
-    def setup_arguments(self):
-        """
-        Set up the command line arguments expected by the program.
-        """
-        # Add command line argument configurations here
-        self.parser.add_argument(
-            "-e", "--extract", help="Initiate the ETL process", action="store_true"
-        )
-        # Add other arguments as needed
-
-    def parse_arguments(self):
-        """
-        Parse and handle the command line arguments.
-        """
-        args = self.parser.parse_args()
-
-        if args.extract:
-            self.execute_etl_process()
-        # Add more argument handling as necessary
-
-    def execute_etl_process(self):
-        """
-        Executes the ETL process by interacting with the ApplicationServer.
-        """
-        self.app_server.process_etl_request()
-        # Add additional logic as necessary
-
-    def display_results(self, results):
-        """
-        Display the results of the ETL process or other commands in a user-friendly manner.
-        """
-        print(results)  # Placeholder for demonstration; implement as needed
-
-
-def main():
-    controller = CLIController()
-    controller.parse_arguments()
-
-
-if __name__ == "__main__":
-    main()
