--- src/cli/cli_utils.py
@@ -1,15 +1,0 @@
-"""Utility functions for CLI operations."""
-
-
-def format_output(data):
-    """
-    Format the given data into a string that is suitable for CLI output.
-
-    Args:
-        data (Any): The data to format.
-
-    Returns:
-        str: The formatted string.
-    """
-    # Implement your formatting logic here
-    return str(data)
