--- src/core/application_server.py
@@ -1,21 +1,0 @@
-"""This module contains the ApplicationServer class, which coordinates the ETL process and handles requests from the CLI."""
-
-
-class ApplicationServer:
-    """
-    The ApplicationServer acts as the central coordinator for initiating and
-    managing the ETL process, and serves as the intermediary between the CLI
-    and the core ETL components.
-    """
-
-    def process_etl_request(self):
-        """Initiates the ETL process."""
-        pass
-
-    def get_results(self):
-        """Retrieves the results of the ETL process."""
-        pass
-
-    def get_visualization_data(self):
-        """Prepares data for visualization after ETL process."""
-        pass
