--- src/services/data_visualization_service.py
@@ -1,16 +1,0 @@
-"""This module contains the DataVisualizationService class for generating visualizations from analysis results."""
-
-
-class DataVisualizationService:
-    """
-    The DataVisualizationService handles the creation of visual representations
-    for the data processed by the application.
-    """
-
-    def generate_graph_visualizations(self, graph_data):
-        """Generates a visual representation of the graph."""
-        pass
-
-    def generate_statistical_charts(self, data):
-        """Generates statistical charts from the given data."""
-        pass
