--- src/services/graph_analysis_service.py
@@ -1,20 +1,0 @@
-"""This module contains the GraphAnalysisService class for performing graph analysis operations."""
-
-
-class GraphAnalysisService:
-    """
-    The GraphAnalysisService provides various analytical operations on the graph data,
-    such as centrality calculation, community detection, and pathfinding.
-    """
-
-    def calculate_centralities(self):
-        """Calculates various centrality metrics for the nodes in the graph."""
-        pass
-
-    def detect_communities(self):
-        """Detects communities within the graph."""
-        pass
-
-    def find_shortest_paths(self, source, target):
-        """Finds the shortest path between two nodes."""
-        pass
