--- src/utils/config.py
@@ -1,10 +1,0 @@
-"""This module manages configuration settings and environment management for the application."""
-
-
-class Config:
-    """
-    The Config class is responsible for managing configuration settings,
-    including environment-specific settings and constants.
-    """
-
-    # Define your configuration settings and methods here
