--- src/utils/helpers.py
@@ -1,9 +1,0 @@
-"""This module contains miscellaneous helper functions used across the application."""
-
-
-def helper_function():
-    """
-    An example helper function that can be used throughout the application
-    to perform a common task.
-    """
-    pass
