+++ trade_network/diagrama_de_componentes.puml
@@ -1,0 +1,31 @@
+@startuml
+
+'TODO: cambiar todo a español y también el código'
+
+package "Análisis de Comercio Internacional" {
+    [ClassificationScheme] -up-> [TradeNetwork] : "Filtra"
+
+    [EconomicComplexityAnalyzer] --> [TradeNetwork] : "Calcula para cada año"
+    [EconomicComplexityAnalyzer] --> [DiversityCalculator] : "Analisis de Flujo"
+
+    package "Clasificación" {
+        [ClassificationScheme] <|-- [SinClasificación]
+        [ClassificationScheme] <|-- [byRegion]
+        [ClassificationScheme] <|-- [byGlobalSouthAndNorth]
+        [ClassificationScheme] <|-- [byIncomeLevel]
+    }
+
+    [TradeNetwork] --> [DataFrame] : "Usa"
+    [DiversityCalculator] --> [stats] : "Calcula entropía"
+    [EconomicComplexityAnalyzer] --> [Parallel] : "Procesamiento paralelo"
+    [EconomicComplexityAnalyzer] --> [DataFrame] : "Almacena resultados"
+
+}
+
+package "Librerías Externas" {
+    [Pandas::DataFrame]
+    [SciPy::stats]
+    [Joblib::Parallel]
+}
+
+@enduml
