+++ trade_network/requirements.txt
@@ -1,0 +1,7 @@
+joblib==1.4.2
+networkx==3.3
+neo4j==5.20.0
+numpy==1.26.4
+pandas==2.2.2
+scipy==1.13.0
+tqdm==4.66.4
